updates from main rig;
 Conflicts:
	.idea/workspace.xml

+++ db/access/datamodels/sql/CREATETABLE/AnnualCarryoverEstimates.sql
@@ -18,9 +18,9 @@ CREATE TABLE AnnualCarryoverEstimates
     ProgramAreaName TEXT(80) NULL DEFAULT NS,
     BocCode TEXT(80) NULL DEFAULT NS,
     BocName TEXT(80) NULL DEFAULT NS,
-    AvailableBalance DECIMAL NULL DEFAULT 0.0,
-    OpenCommitments DECIMAL NULL DEFAULT 0.0,
-    UnobligatedAuthority DECIMAL NULL DEFAULT 0.0,
+    AvailableBalance DOUBLE NULL DEFAULT 0.0,
+    OpenCommitments DOUBLE NULL DEFAULT 0.0,
+    UnobligatedAuthority DOUBLE NULL DEFAULT 0.0,
     CONSTRAINT AnnualCarryoverEstimatesId
         PRIMARY KEY(AnnualCarryoverEstimatesId)
 );
