+++ db/access/datamodels/sql/CREATETABLE/UnobligatedBalances.sql
@@ -9,7 +9,7 @@ CREATE TABLE UnobligatedBalances
     FundName TEXT(80) NULL DEFAULT NS,
     AccountNumber TEXT(80) NULL DEFAULT NS,
     AccountName TEXT(80) NULL DEFAULT NS,
-    Amount DECIMAL NULL DEFAULT 0.0,
+    Amount DOUBLE NULL DEFAULT 0.0,
     CONSTRAINT UnobligatedBalancePrimaryKey
         PRIMARY KEY(UnobligatedBalancesId)
 );
